Rename: logical_ops_test -> logicalop_test
